+++ src/components/layer/TLayerRouter.vue
@@ -3,13 +3,11 @@
     <Card :pt="cardLayoutPt()">
       <!-- 卡片头插槽 -->
       <template v-if="!hideHeader" #title>
-        <div class="flex items-center gap-2">
+        <div class="flex items-center gap-2 flexSplit">
           <span class="flexVY">
             <span class="font-bold">{{ imt.bizIdentNamezh || '未定义' }}</span>
             <span v-if="imt.description" class="text-surface-500 text-sm font-medium dark:text-surface-400">
-              <slot name="subTitle">
-                {{ imt.description }}
-              </slot>
+              <slot name="subTitle"></slot>
             </span>
           </span>
           <!-- 标题， 增删查改等按钮的插槽 -->
@@ -17,11 +15,6 @@
           </slot>
         </div>
       </template>
-      <!-- <template v-if="!hideIcons" #icons>
-        <slot name="icons"></slot>
-        <Button icon="pi pi-cog" severity="secondary" rounded text />
-        <Menu ref="refOptMenu" id="config_menu" :model="ROUTER_PANEL_ITEMS" popup />
-      </template> -->
 
       <template #content>
         <div class="h-full">
@@ -33,13 +26,6 @@
       <!-- 卡片脚注插槽 -->
       <template v-if="!hideFooter" #footer>
         <slot name="footer">
-          <!-- <div class="flex flex-wrap items-center justify-between gap-4">
-            <div class="flex items-center gap-2">
-              <Button icon="pi pi-user" rounded text></Button>
-              <Button icon="pi pi-bookmark" severity="secondary" rounded text></Button>
-            </div>
-            <span class="text-surface-500 dark:text-surface-400">Updated 2 hours ago</span>
-          </div> -->
         </slot>
       </template>
     </Card>