+++ src/constant/uiKit/index.ts
@@ -6,3 +6,6 @@ export const svgStockeWidth = 1.5;
 export const badgeOverlayColor = 'orange';
 
 export const TABLE_SIZE: TTableSize = 'large';
+
+/** 表单样式-分栏数 */
+export const FORM_COLS: 1 | 2 | 3 | 4 = 2;